background image is customised by category

+++ Volunti/src/main/java/com/codecool/volunti/service/model/OrganisationService.java
@@ -14,6 +14,7 @@ import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.File;
 import java.util.List;
 
 @Slf4j
@@ -56,8 +57,7 @@ public class OrganisationService {
             String oldBackgroundPicture = organisation.getBackgroundPicture();
             String newFileName = storageService.store(organisation.getBackgroundPictureFileForSave());
             organisation.setBackgroundPicture(newFileName);
-            storageService.deleteOne(oldBackgroundPicture);
-
+            if(oldBackgroundPicture != null) storageService.deleteOne(oldBackgroundPicture);
         }
         return organisationRepository.save(organisation);
     }
@@ -109,6 +109,13 @@ public class OrganisationService {
         oldItem.setAddress(newItem.getAddress());
         oldItem.setCity(newItem.getCity());
         oldItem.setZipcode(newItem.getZipcode());
+    }
+
+    public void setDefaultBackgroundImage(Organisation organisation) {
+        String staticPath = "Volunti/src/main/resources/static/images/background_image/";
+        File testBackgroundImageFile = new File(staticPath + organisation.getCategory().name() + ".jpg" );
+        organisation.setBackgroundPictureFileForSave(testBackgroundImageFile);
+        save(organisation);
     }
 
 }
